FLyttet ut tester knyttet til henlegging fra behandlingServiceTest til henleggBehandligServiceTest, flyttet ut dataklasser knyttet til henleggelsesbrev til egen fil
